backend/requirements.txt
fastapi>=0.110
uvicorn>=0.24
gunicorn>=21
pydantic>=2.6
pydantic-settings>=2.2
pydantic-ai>=0.0.14
fastmcp>=0.3.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
httpx>=0.26
tenacity>=8.2
python-jose[cryptography]>=3.3
slowapi>=0.1.9
structlog>=24.1
jsonschema>=4.21
sse-starlette>=1.8
redis>=5.0
passlib[bcrypt]>=1.7
bcrypt==3.2.2
email-validator>=2.1
logfire>=0.46.0
pgvector>=0.2.0
openai>=1.12
openpyxl>=3.1
xlrd>=2.0
websockets>=12.0
jmespath>=1.0